requirements.txt
click==7.1.2
cycler==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
nltk==3.5
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
regex==2020.11.13
six==1.15.0
tqdm==4.56.0
